requirements.txt
# ==========================================================
# OrangeCashMachine – Market Data / Orchestration Dependencies
# Principios aplicados: SOLID, KISS, DRY, SafeOps
# ==========================================================

# Orquestación y flujo de tareas
prefect==2.19.5         # Prefect 2.x para flows y UI
griffe==0.32.3          # Requerido por Prefect 2.x para introspección de docstrings

# Validación y tipado
pydantic==2.8.2         # Modelos de datos robustos
pandera==0.19.3         # Validación de DataFrames de Pandas

# Procesamiento de datos
numpy<2.0               # Evita breaking change np.string_
pandas==2.2.2           # Procesamiento de tablas y series temporales
pyarrow==16.1.0         # Serialización y parquet

# Conectividad con exchanges
ccxt==4.3.58            # Conexión a exchanges para datos de mercado

# Logging y observabilidad
loguru==0.7.2           # Logging estructurado y seguro

# Configuración y archivos
pyyaml==6.0.2           # Lectura y parse de YAML

# Resiliencia
pybreaker==1.1.0        # Circuit breaker para llamadas externas